refactor: extract shared vault logic into _shared.ts and align assessPoolRisk naming
